requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
asgiref==3.6.0
openai==1.30.3
selenium==4.9.1
g4f==0.3.1.9
g4f[all]
undetected-chromedriver==3.4.7
boto3==1.26.152
psycopg2-binary==2.9.9